Revert "SM-3949: Fixed regex issue, Added custom hooks for tqm"
