+++ web/micro-ui-internals/packages/libraries/src/hooks/index.js
@@ -244,7 +244,6 @@ import useBillAmendmentInbox from "./billAmendment/useInbox";
 import { usePrivacyContext } from "./usePrivacyContext";
 import useToCheckPrivacyEnablement from "./ws/useToCheckPrivacyEnablement";
 import useWSConfigMDMS from "./ws/useWSConfigMDMS";
-import useCustomAPITqmHook from "./useCustomApiTqmHook";
 
 const pgr = {
   useComplaintDetails,
@@ -545,7 +544,6 @@ const Hooks = {
   useBulkPdfDetails,
   useBillAmendmentInbox,
   useAudit,
-  useCustomAPITqmHook,
 };
 
 export default Hooks;
