+++ web/micro-ui-internals/packages/libraries/src/hooks/useCustomAPIHook.js
@@ -2,30 +2,53 @@ import { useQuery, useQueryClient } from "react-query";
 import { CustomService } from "../services/elements/CustomService";
 
 /**
- * Custom hook which can gives the privacy functions to access
+ * Custom hook which can make api call and format response
  *
  * @author jagankumar-egov
  *
- * Feature :: Privacy
  *
  * @example
- *         const { privacy , updatePrivacy } = Digit.Hooks.usePrivacyContext()
+ * 
+ const requestCriteria = [
+      "/user/_search",             // API details
+    {},                            //requestParam
+    {data : {uuid:[Useruuid]}},    // requestBody
+    {} ,                           // privacy value 
+    {                              // other configs
+      enabled: privacyState,
+      cacheTime: 100,
+      select: (data) => {
+                                    // format data
+        return  _.get(data, loadData?.jsonPath, value);
+      },
+    },
+  ];
+  const { isLoading, data, revalidate } = Digit.Hooks.useCustomAPIHook(...requestCriteria);
+
  *
- * @returns {Object} Returns the object which contains privacy value and updatePrivacy method
+ * @returns {Object} Returns the object which contains data and isLoading flag
  */
-const useCustomAPIHook = (url, params, body, plainAccessRequest, options = {}) => {
+
+const useCustomAPIHook = (props) => {
+  const { url = "", method, params, body, config = {}, plainAccessRequest, changeQueryName = "Random" } = props;
+
   const client = useQueryClient();
-  //api name, querystr, reqbody
-  const { isLoading, data } = useQuery(
-    ["CUSTOM", { ...params, ...body, ...plainAccessRequest }].filter((e) => e),
-    () => CustomService.getResponse({ url, params, ...body, plainAccessRequest }),
-    options
+
+  const { isLoading, data, isFetching } = useQuery(
+    [url, changeQueryName].filter((e) => e),
+    () => CustomService.getResponse({ url, params, method, body, plainAccessRequest }),
+    {
+      cacheTime: 0,
+      ...config,
+    }
   );
+
   return {
     isLoading,
+    isFetching,
     data,
     revalidate: () => {
-      data && client.invalidateQueries({ queryKey: ["CUSTOM", { ...params, ...body, ...plainAccessRequest }] });
+      data && client.invalidateQueries({ queryKey: [url]?.filter((e) => e) });
     },
   };
 };
