+++ web/micro-ui-internals/packages/modules/fsm/src/pages/employee/FstpAddVehicle.js
@@ -23,7 +23,7 @@ const FstpAddVehicle = ({ onSelect }) => {
         name: "vehicleNumber",
         validation: {
           // "pattern": "[a-zA-Z0-9 ]{1,20}",
-          pattern: `[A-Z]{2}\\s{1}[0-9]{2}\\s{0,1}[A-Z]{1,2}\\s{1}[0-9]{4}`,
+          pattern: `[A-Z]{2}[0-9]{2}[A-Z]{0,2}[0-9]{4}`,
           title: t("ES_FSM_VEHICLE_FORMAT_TIP"),
         },
       },
